+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Breaking**: generic: `&ObjectHandle`を`ObjectHandle`に変更
 - module: FromScriptModuleParamTableを実装する型を追加
+- traitの型をRust Analyzerが扱いやすくなるように変更
 
 ### デモプラグイン